LayoutTests/fast/profiler/two-execution-contexts.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("Two Execution Contexts");
function testStart()
{
    intermediaryFunction();
}

function intermediaryFunction()
{
    testEnd();
}
</script>
</head>

<body onload="testStart()">
This page has two script tags with some simple JavaScript.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be two (program) elements.
</body>
<script>
function testEnd() {
    console.profileEnd();
    printProfilesDataWithoutTime();
}
</script>
</html>
