LayoutTests/fast/profiler/two-execution-contexts-expected.txt

CONSOLE MESSAGE: line 59: Two Execution Contexts
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) two-execution-contexts.html 8
CONSOLE MESSAGE: line 75:   (program) two-execution-contexts.html 4
CONSOLE MESSAGE: line 75:   (program) two-execution-contexts.html 28
CONSOLE MESSAGE: line 75:   onload two-execution-contexts.html 21
CONSOLE MESSAGE: line 75:    testStart two-execution-contexts.html 10
CONSOLE MESSAGE: line 75:     intermediaryFunction two-execution-contexts.html 15
CONSOLE MESSAGE: line 75:      testEnd two-execution-contexts.html 29
CONSOLE MESSAGE: line 75:  (idle)  0
This page has two script tags with some simple JavaScript. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be two (program) elements.
